issue #7: tests to demonstrate the issue
First, there is a test to show that `elli_fileserve:local_path` fails
to contain the result within the root `path`.

It requires that the requested path is absolute, rather than
relative. Hence, a set of tests to show that we indeed can get an
absolute path for arbitrary settings of `prefix`.

Next commit will address these issues. I've kept them separate to make
it easier to verify that the tests indeed point at an issue, and that
the following commit does indeed solve them.