fixes #7
Make it the responsibility for `elli_fileserver:local_path/2` to make
sure it only process relative paths (see note below).

Drop the responsibility to remove any leading `/` from
`elli_fileserve:unprefix/2` (as it wasn't doing a very good job at it
any way). `unprefix_test_` adapted for this change.

Note: I think the culprit to this issue may be the oversight of this
little piece of information, from the documentation for
`filename:join/1`::

     Joins a list of file name Components with directory
     separators. If one of the elements of Components includes an
     absolute path, for example "/xxx", the preceding elements, if
     any, are removed from the result.